pyproject.toml
[project]
name = "bblocks-datacommons-tools"
version = "0.0.1"
description = "Tools to work with Data Commons. Part of the bblocks projects."
authors = [
    {name = "ONE Campaign"},
    {name = "Luca Picci",email = "lpicci96@gmail.com"},
    {name = "Jorge Rivera", email = "jorge.rivera@one.org"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic (>=2.11.3,<3.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "google-cloud-storage (>=3.1.0,<4.0.0)",
    "google-cloud-run (>=0.10.17,<0.11.0)",
]

[tool.poetry]
packages = [{include = "bblocks", from = "src"}]


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
